New Platforms API based toolchain configuration.
Instead of selecting which make/cmake toolchain to based on flags, now
we select one based on the target CPU, which can be wasm32 or wasm64.
This will register an Emscripten-based C/C++ toolchain suite.

Now all libraries should be buildable for the host:

```
$ bazel build //lib/...
```

And for WebAssembly:

```
$ bazel build --config=wasm64 //libl/...
```

NOTE: --config=wasm{32,64} and -c opt currently don't play nicely.