Add a hacky workaround for not founding the `emar` (static) linker.
This is because Bazel seems to prefix the tool_path binaries with
toolchains/ when they seem like a relative path, and then CMake doesn't
find them because the environment variables (and CMake flags) are not
constructed well.

This workaround adds some simple shell scripts that forward the call to
where the emcc/em++/emar files actually live.