Use wasm32 Starlark transition instead of wasm64.
Before we never specified MEMORY64=1, so the default was used
(MEMORY64=0), which implies that the underlying architecture is actually
wasm64. This change makes sure that both --config=wasm32 and
--config=wasm64 would pass the right compile flags.

Note that wasm_library() currently defaults to wasm32 as that is what we
were building anyway. A follow-up change might split into wasm32 and
wasm64 targets, if it turns out that libraries would actually build with
MEMORY64=1.