Set CMAKE_TRY_COMPILE_TARGET_TYPE to STATIC_LIBRARY.
This works around an issue where CMake fails to link binary files.

This is probably not a proper solution, but good enough to build static
libraries, which is what we want to build with CMake anyway: the actual
binaries are compiled using the wasm_library() rule (for now).

See https://github.com/bazelbuild/rules_foreign_cc/pull/523 for details.