Update rules_foreign_cc to 0.3.0.
This change removes the `emcmake` and `emmake` prefixes, instead plain
`cmake` and `make` is called by Bazel, but with additional environment
variables to enable cross-compilation.

Example tests all pass:

$ bazel test //examples/...

GRASS fails to build as it fails to find PROJ4, but other libs seem to
build at least for k8. Further tests for wasm32 builds need to be added
to ensure the WebAssembly builds are not broken by similar updates.