Bump Emscripten to 2.0.24.
This fixes a compilation error with //lib/openssl, as well as allows
using remote dependencies for the Emscripten SDK instead of listing the
repos explicitly. (This allows for easier Emscripten updates in the
future.)

Fix some tests and builds now that apparently the new Emscripten/Clang
combo has extra warnings. Also, no need to patch the Emscripten SDK any
more.