Make tool_paths relative.
This requires moving the toolchain configuration to the workspace root
(`//`) since by default toolchain paths are prefixed with the package
where the toolchain is defined.

In theory, this can be changed by setting tool_path_origin to
WORKSPACE_ROOT (the default is CROSSTOOL_PACKAGE for legacy reasons),
however I'm not sure how to do that with the Starlark macros I'm
currently using.

This should not change any foreign CC rules, but should allow the CC
toolchain to be used with cc_library and cc_binary rules too.